refactor: (#17) Authority enum으로 변경
