+++ src/infrastructure/domain/notification/presentation/notification.web.adapter.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Post, Get, HttpCode, Patch, Query, Param } from '@nestjs/common';
 import { SetDeviceTokenWebRequest } from './dto/notification.web.dto';
 import { Permission } from '../../../global/decorator/authority.decorator';
-import { Authority } from '../../user/persistence/user.entity';
+import { Authority } from '../../../../application/domain/user/authority';
 import { CurrentUser } from '../../../global/decorator/current-user.decorator';
 import { User } from '../../../../application/domain/user/user';
 import { SetDeviceTokenUseCase } from '../../../../application/domain/notification/usecase/set-device-token.usecase';
