+++ src/infrastructure/domain/suggestion/presentation/suggestion.web.adapter.ts
@@ -3,17 +3,17 @@ import { CreateSuggestionUseCase } from '../../../../application/domain/suggesti
 import { UpdateSuggestionUseCase } from '../../../../application/domain/suggestion/usecase/update-suggestion.usecase';
 import { DeleteSuggestionUseCase } from '../../../../application/domain/suggestion/usecase/delete-suggestion.usecase';
 import { QueryMySuggestionsUseCase } from '../../../../application/domain/suggestion/usecase/query-my-suggestions.usecase';
-import { Permission } from 'src/infrastructure/global/decorator/authority.decorator';
-import { Authority } from '../../user/persistence/user.entity';
-import { CurrentUser } from 'src/infrastructure/global/decorator/current-user.decorator';
+import { Permission } from '../../../global/decorator/authority.decorator';
+import { Authority } from '../../../../application/domain/user/authority';
+import { CurrentUser } from '../../../global/decorator/current-user.decorator';
 import {
     CreateSuggestionResponse,
     QueryAllSuggestionsResponse,
     QueryMySuggestionsResponse
-} from 'src/application/domain/suggestion/dto/suggestion.dto';
+} from '../../../../application/domain/suggestion/dto/suggestion.dto';
 import { User } from '../../../../application/domain/user/user';
 import { SuggestionWebRequest } from './dto/suggestion.web.dto';
-import { QueryAllSuggestionsUseCase } from 'src/application/domain/suggestion/usecase/query-all-suggestions.usecase';
+import { QueryAllSuggestionsUseCase } from '../../../../application/domain/suggestion/usecase/query-all-suggestions.usecase';
 
 @Controller('suggestions')
 export class SuggestionWebAdapter {
